Split boofuzz/sessions.py
