+++ boofuzz/sessions/__init__.py
@@ -1,0 +1,7 @@
+from .connection import Connection
+from .session_info import SessionInfo
+from .session import Session, open_test_run
+from .target import Target
+from .web_app import WebApp
+
+__all__ = [ Connection, SessionInfo, Target, Session, WebApp, open_test_run ]
