+++ boofuzz/sessions/connection.py
@@ -1,0 +1,29 @@
+from ..pgraph import Edge
+
+class Connection(Edge):
+    def __init__(self, src, dst, callback=None):
+        """
+        Extends pgraph.edge with a callback option. This allows us to register a function to call between node
+        transmissions to implement functionality such as challenge response systems. The callback method must follow
+        this prototype::
+
+            def callback(target, fuzz_data_logger, session, node, edge, *args, **kwargs)
+
+        Where node is the node about to be sent, edge is the last edge along the current fuzz path to "node", session
+        is a pointer to the session instance which is useful for snagging data such as sesson.last_recv which contains
+        the data returned from the last socket transmission and sock is the live socket. A callback is also useful in
+        situations where, for example, the size of the next packet is specified in the first packet.
+
+        Args:
+            src (int): Edge source ID
+            dst (int): Edge destination ID
+            callback (function): Optional. Callback function to pass received data to between node xmits
+
+        .. versionchanged:: 0.5.0
+            This class has been moved into the sessions subpackage. The full path is now boofuzz.sessions.connection.Connection.
+
+        """
+
+        super(Connection, self).__init__(src, dst)
+
+        self.callback = callback
