+++ boofuzz/sessions/session.py
@@ -13,10 +13,6 @@ import zlib
 from builtins import input
 from io import open
 
-from tornado.httpserver import HTTPServer
-from tornado.ioloop import IOLoop
-from tornado.wsgi import WSGIContainer
-
 from boofuzz import (
     blocks,
     constants,
@@ -33,337 +29,19 @@ from boofuzz import (
 from boofuzz.monitors import CallbackMonitor
 from boofuzz.mutation_context import MutationContext
 from boofuzz.protocol_session import ProtocolSession
+
+from ..exception import BoofuzzFailure
+
+from .connection import Connection
+from .session_info import SessionInfo
+from .web_app import WebApp
+
+# needed for build_web_app_thread
+from tornado.httpserver import HTTPServer
+from tornado.ioloop import IOLoop
+from tornado.wsgi import WSGIContainer
+
 from boofuzz.web.app import app
-from .exception import BoofuzzFailure
-
-
-class Target:
-    """Target descriptor container.
-
-    Takes an ITargetConnection and wraps send/recv with appropriate
-    FuzzDataLogger calls.
-
-    Encapsulates pedrpc connection logic.
-
-    Contains a logger which is configured by Session.add_target().
-
-    Example:
-        tcp_target = Target(SocketConnection(host='127.0.0.1', port=17971))
-
-    Args:
-        connection (itarget_connection.ITargetConnection): Connection to system under test.
-        monitors (List[Union[IMonitor, pedrpc.Client]]): List of Monitors for this Target.
-        monitor_alive: List of Functions that are called when a Monitor is alive. It is passed
-                          the monitor instance that became alive. Use it to e.g. set options
-                          on restart.
-        repeater (repeater.Repeater): Repeater to use for sending. Default None.
-        procmon: Deprecated interface for adding a process monitor.
-        procmon_options: Deprecated interface for adding a process monitor.
-
-    """
-
-    def __init__(
-        self,
-        connection,
-        monitors=None,
-        monitor_alive=None,
-        max_recv_bytes=10000,
-        repeater=None,
-        procmon=None,
-        procmon_options=None,
-        **kwargs
-    ):
-        self._fuzz_data_logger = None
-
-        self._target_connection = connection
-        self.max_recv_bytes = max_recv_bytes
-        self.repeater = repeater
-        self.monitors = monitors if monitors is not None else []
-        if procmon is not None:
-            if procmon_options is not None:
-                procmon.set_options(**procmon_options)
-            self.monitors.append(procmon)
-
-        self.monitor_alive = monitor_alive if monitor_alive is not None else []
-
-        if "procmon" in kwargs.keys() and kwargs["procmon"] is not None:
-            warnings.warn(
-                "Target(procmon=...) is deprecated. Please change your code"
-                " and add it to the monitors argument. For now, we do this "
-                "for you, but this will be removed in the future.",
-                FutureWarning,
-            )
-            self.monitors.append(kwargs["procmon"])
-
-        if "netmon" in kwargs.keys() and kwargs["netmon"] is not None:
-            warnings.warn(
-                "Target(netmon=...) is deprecated. Please change your code"
-                " and add it to the monitors argument. For now, we do this "
-                "for you, but this will be removed in the future.",
-                FutureWarning,
-            )
-            self.monitors.append(kwargs["netmon"])
-
-        # set these manually once target is instantiated.
-        self.vmcontrol = None
-        self.vmcontrol_options = {}
-
-    @property
-    def netmon_options(self):
-        raise NotImplementedError(
-            "This property is not supported; grab netmon from monitors and use set_options(**dict)"
-        )
-
-    @property
-    def procmon_options(self):
-        raise NotImplementedError(
-            "This property is not supported; grab procmon from monitors and use set_options(**dict)"
-        )
-
-    def close(self):
-        """
-        Close connection to the target.
-
-        :return: None
-        """
-        self._fuzz_data_logger.log_info("Closing target connection...")
-        self._target_connection.close()
-        self._fuzz_data_logger.log_info("Connection closed.")
-
-    def open(self):
-        """
-        Opens connection to the target. Make sure to call close!
-
-        :return: None
-        """
-        self._fuzz_data_logger.log_info("Opening target connection ({0})...".format(self._target_connection.info))
-        self._target_connection.open()
-        self._fuzz_data_logger.log_info("Connection opened.")
-
-    def pedrpc_connect(self):
-        warnings.warn(
-            "pedrpc_connect has been renamed to monitors_alive. "
-            "This alias will stop working in a future version of boofuzz.",
-            FutureWarning,
-        )
-
-        return self.monitors_alive()
-
-    def monitors_alive(self):
-        """
-        Wait for the monitors to become alive / establish connection to the RPC server.
-        This method is called on every restart of the target and when it's added to a session.
-        After successful probing, a callback is called, passing the monitor.
-
-        :return: None
-        """
-        for monitor in self.monitors:
-            while True:
-                if monitor.alive():
-                    break
-                time.sleep(1)
-
-            if self.monitor_alive:
-                for cb in self.monitor_alive:
-                    cb(monitor)
-
-    def recv(self, max_bytes=None):
-        """
-        Receive up to max_bytes data from the target.
-
-        Args:
-            max_bytes (int): Maximum number of bytes to receive.
-
-        Returns:
-            Received data.
-        """
-        if max_bytes is None:
-            max_bytes = self.max_recv_bytes
-
-        if self._fuzz_data_logger is not None:
-            self._fuzz_data_logger.log_info("Receiving...")
-
-        data = self._target_connection.recv(max_bytes=max_bytes)
-
-        if self._fuzz_data_logger is not None:
-            self._fuzz_data_logger.log_recv(data)
-
-        return data
-
-    def send(self, data):
-        """
-        Send data to the target. Only valid after calling open!
-
-        Args:
-            data: Data to send.
-
-        Returns:
-            None
-        """
-        num_sent = 0
-        if self._fuzz_data_logger is not None:
-            repeat = ""
-            if self.repeater is not None:
-                repeat = ", " + self.repeater.log_message()
-
-            self._fuzz_data_logger.log_info("Sending {0} bytes{1}...".format(len(data), repeat))
-
-        if self.repeater is not None:
-            self.repeater.start()
-            while self.repeater.repeat():
-                num_sent = self._target_connection.send(data=data)
-            self.repeater.reset()
-        else:
-            num_sent = self._target_connection.send(data=data)
-
-        if self._fuzz_data_logger is not None:
-            self._fuzz_data_logger.log_send(data[:num_sent])
-
-    def set_fuzz_data_logger(self, fuzz_data_logger):
-        """
-        Set this object's fuzz data logger -- for sent and received fuzz data.
-
-        :param fuzz_data_logger: New logger.
-        :type fuzz_data_logger: ifuzz_logger.IFuzzLogger
-
-        :return: None
-        """
-        self._fuzz_data_logger = fuzz_data_logger
-
-
-class Connection(pgraph.Edge):
-    def __init__(self, src, dst, callback=None):
-        """
-        Extends pgraph.edge with a callback option. This allows us to register a function to call between node
-        transmissions to implement functionality such as challenge response systems. The callback method must follow
-        this prototype::
-
-            def callback(target, fuzz_data_logger, session, node, edge, *args, **kwargs)
-
-        Where node is the node about to be sent, edge is the last edge along the current fuzz path to "node", session
-        is a pointer to the session instance which is useful for snagging data such as sesson.last_recv which contains
-        the data returned from the last socket transmission and sock is the live socket. A callback is also useful in
-        situations where, for example, the size of the next packet is specified in the first packet.
-
-        Args:
-            src (int): Edge source ID
-            dst (int): Edge destination ID
-            callback (function): Optional. Callback function to pass received data to between node xmits
-        """
-
-        super(Connection, self).__init__(src, dst)
-
-        self.callback = callback
-
-
-class SessionInfo:
-    def __init__(self, db_filename):
-        self._db_reader = fuzz_logger_db.FuzzLoggerDbReader(db_filename=db_filename)
-
-    @property
-    def monitor_results(self):
-        return self._db_reader.failure_map
-
-    @property
-    def monitor_data(self):
-        return {-1, "Monitor Data is not currently saved in the database"}
-
-    @property
-    def procmon_results(self):
-        warnings.warn(
-            "procmon_results has been renamed to monitor_results."
-            "This alias will stop working in a future version of boofuzz",
-            FutureWarning,
-        )
-        return self.monitor_results
-
-    @property
-    def netmon_results(self):
-        warnings.warn(
-            "netmon_results is now part of monitor_data" "This alias will stop working in a future version of boofuzz",
-            FutureWarning,
-        )
-        return self.monitor_data
-
-    @property
-    def fuzz_node(self):
-        return None
-
-    @property
-    def total_num_mutations(self):
-        return None
-
-    @property
-    def total_mutant_index(self):
-        x = next(self._db_reader.query("SELECT COUNT(*) FROM cases"))[0]
-        return x
-
-    @property
-    def mutant_index(self):
-        return None
-
-    def test_case_data(self, index):
-        """Return test case data object (for use by web server)
-
-        Args:
-            index (int): Test case index
-
-        Returns:
-            Test case data object
-        """
-        return self._db_reader.get_test_case_data(index=index)
-
-    @property
-    def is_paused(self):
-        return False
-
-    @property
-    def state(self):
-        return "finished"
-
-    @property
-    def exec_speed(self):
-        return 0
-
-    @property
-    def runtime(self):
-        return 0
-
-    @property
-    def current_test_case_name(self):
-        return ""
-
-
-class WebApp:
-    """Serve fuzz data over HTTP.
-
-    Args:
-        session_info (SessionInfo): Object providing information on session
-        web_port (int):         Port for monitoring fuzzing campaign via a web browser. Default 26000.
-        web_address (string):   Address binded to port for monitoring fuzzing campaign via a web browser.
-                                Default 'localhost'.
-    """
-
-    def __init__(
-        self, session_info, web_port=constants.DEFAULT_WEB_UI_PORT, web_address=constants.DEFAULT_WEB_UI_ADDRESS
-    ):
-        self._session_info = session_info
-        self._web_interface_thread = self._build_webapp_thread(port=web_port, address=web_address)
-        pass
-
-    def _build_webapp_thread(self, port, address):
-        app.session = self._session_info
-        http_server = HTTPServer(WSGIContainer(app))
-        http_server.listen(port, address=address)
-        flask_thread = threading.Thread(target=IOLoop.instance().start)
-        flask_thread.daemon = True
-        return flask_thread
-
-    def server_init(self):
-        """Called by fuzz() to initialize variables, web interface, etc."""
-        if not self._web_interface_thread.is_alive():
-            # spawn the web interface.
-            self._web_interface_thread.start()
 
 
 def open_test_run(db_filename, port=constants.DEFAULT_WEB_UI_PORT, address=constants.DEFAULT_WEB_UI_ADDRESS):
@@ -427,6 +105,9 @@ class Session(pgraph.Graph):
         db_filename (str):      Filename to store sqlite db for test results and case information.
                                 Defaults to ./boofuzz-results/{uniq_timestamp}.db
         web_address:            Address where's Boofuzz logger exposed. Default 'localhost'
+
+    .. versionchanged:: 0.5.0
+        This class has been moved into the sessions subpackage. The full path is now boofuzz.sessions.session.Session.
     """
 
     def __init__(
